% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_velocity.R
\name{calculate_velocity}
\alias{calculate_velocity}
\title{Calculate Velocity}
\format{
Input data frame columns
\describe{
  \item{gazeHasValue}{Logical (boolean) if there is valid gaze data}
  \item{eyeDataRelativeTimestamp}{Timestamp of the data}
  \item{gazeorigin_x}{X coordinates of the gaze origin}
  \item{gazeorigin_y}{Y coordinates of the gaze origin}
  \item{gazeorigin_z}{Z coordinates of the gaze origin}
  \item{gazePoint_x}{X coordinates of the gaze point}
  \item{gazePoint_y}{Y coordinates of the gaze point}
  \item{gazePoint_z}{Z coordinates of the gaze point}
}
}
\usage{
calculate_velocity(data, window_length = 20)
}
\arguments{
\item{data}{Data frame of the eye tracking data we want to process}

\item{window_length}{Length of the window over which we want to calculate the velocity}
}
\value{
The input data frame with the additional column \emph{velocity}
}
\description{
Calculate the velocity of each gaze point (results required for I-VT fixation filter)
}
\details{
This is an implementation of the velocity calculation as documented in the \href{https://www.tobiipro.com/siteassets/tobii-pro/learn-and-support/analyze/how-do-we-classify-eye-movements/tobii-pro-i-vt-fixation-filter.pdf}{Tobii Pro I-VT fixation filter}.
It first calculates the window size (number of data points) over which the velocity should be calculated based on the specified window length (in ms)
and the average time between data points of the first 100 samples. This accommodates data logs with non-uniform time between data points.
Then it calculates the velocity for each gaze point by calculating the angle between the first and last gaze point in the current window
with the origin of the gaze point in the middle of the window. This angle divided by the window length is saved as the velocity of the middle gaze point.

The default window length of 20ms and the eye tracker rate of the Microsoft HoloLens 2 of 30Hz results in a window size of two and therefore a
calculation of the velocity between the previous and current gaze point.
}
