% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_iaoi.R
\name{classify_iaoi}
\alias{classify_iaoi}
\title{Classify Fixations by AOI}
\format{
Input data frame columns
\describe{
  \item{eyeDataRelativeTimestamp}{Timestamp of the data}
  \item{gazePoint_x}{X coordinates of the gaze point}
  \item{gazePoint_y}{Y coordinates of the gaze point}
  \item{gazePoint_z}{Z coordinates of the gaze point}
  \item{gazePointAOI_name}{Name of the AOI which was hit}
}

Additional columns in output
\describe{
  \item{eventIndex}{Index of the event in which the gaze point lies}
  \item{eventType}{Type of the current event}
  \item{eventDuration}{Duration of the current event}
  \item{fixation_x}{X coordinate of the current fixation (if eventType is a fixation)}
  \item{fixation_y}{Y coordinate of the current fixation (if eventType is a fixation)}
  \item{fixation_z}{Z coordinate of the current fixation (if eventType is a fixation)}
}
}
\usage{
classify_iaoi(data, min_fixation_duration = 100)
}
\arguments{
\item{data}{Data frame of the eye tracking data we want to process}

\item{min_fixation_duration}{Minimal fixation duration}
}
\value{
The input data frame with the classification columns added
}
\description{
Classify each gaze point based on whether it hits an AOI and a minimum fixation duration
}
\details{
This is a custom implementation of an fixation identification by area of interest.
It first classifies all gaze points which hit an AOI as fixation and all other gaze points as saccades.
Then it groups consecutive gaze points on the same AOI as fixation when the total duration of gaze points exceed the minimum fixation duration.
If the minimum fixation duration isn't exceeded, the gaze points are reclassified as saccades.
}
