% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_ivt.R
\name{classify_ivt}
\alias{classify_ivt}
\title{Classify Fixations by Velocity Threshold}
\format{
Input data frame columns
\describe{
  \item{velocity}{Velocity of the current gaze point}
}

Additional columns in output
\describe{
  \item{classification}{Classification of this gaze point}
  \item{eventIndex}{Index of the event in which the gaze point lies}
  \item{eventType}{Type of the current event}
  \item{eventDuration}{Duration of the current event}
  \item{fixation_x}{X coordinate of the current fixation (if eventType is a fixation)}
  \item{fixation_y}{Y coordinate of the current fixation (if eventType is a fixation)}
  \item{fixation_z}{Z coordinate of the current fixation (if eventType is a fixation)}
}
}
\usage{
classify_ivt(data, velocity_threshold = 100)
}
\arguments{
\item{data}{Data frame of the eye tracking data we want to process}

\item{velocity_threshold}{Velocity threshold in degrees per second (°/s) over which the gaze point should be classified as saccade}
}
\value{
The input data frame with the classification columns added
}
\description{
Classify each gaze point based on their velocities as calculated by \code{\link[arett]{calculate_velocity}}.
}
\details{
This is an implementation of the fixation identification by velocity threshold (I-VT) as documented in the \href{https://www.tobiipro.com/siteassets/tobii-pro/learn-and-support/analyze/how-do-we-classify-eye-movements/tobii-pro-i-vt-fixation-filter.pdf}{Tobii Pro I-VT fixation filter}.
It compares the velocity of each gaze point against the threshold and classifies the gaze point as saccade if the velocity is equal or above it.
Additionally it groups gaze points of the same classification into an event.

The default velocity threshold is 100 degrees per second, as described by Tobii for their mobile eye trackers.
For stationary eye trackers the default threshold by Tobii is 30 degrees per second.
}
