% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard_short_fixations.R
\name{discard_short_fixations}
\alias{discard_short_fixations}
\title{Discard Short Fixations}
\format{
Input data frame columns
\describe{
  \item{eventIndex}{Index of the event in which the gaze point lies}
  \item{eventType}{Type of the current event}
  \item{eventDuration}{Duration of the current event}
  \item{fixation_x}{X coordinate of the current fixation}
  \item{fixation_y}{Y coordinate of the current fixation}
  \item{fixation_z}{Z coordinate of the current fixation}
}
}
\usage{
discard_short_fixations(data, min_duration = 60, reclassify_saccade = FALSE)
}
\arguments{
\item{data}{Data frame of the eye tracking data we want to process}

\item{min_duration}{Minimum duration of a fixation}

\item{reclassify_saccade}{Reclassify discarded fixations as saccade instead of a gap}
}
\value{
The input data frame with the fixation classifications which are too short discarded
}
\description{
Discard fixations which have a duration shorter than the specified minimum duration.
}
\details{
This function removes all fixations which are shorter than the specified minimum duration and reclassifies them as gaps.
If \code{reclassify_saccade} is specified they are classified as saccade instead of a gap.
This post processing is only necessary after the I-VT filter as both I-DT as well as I-AOI inherently include a minimum fixation duration.
}
