% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_fill.R
\name{gap_fill}
\alias{gap_fill}
\title{Fill Gaps}
\format{
Input data frame columns
\describe{
  \item{gazeHasValue}{Boolean if there is valid gaze data}
  \item{eyeDataRelativeTimestamp}{Timestamp of the data}
  \item{gazeorigin_x}{X coordinates of the gaze origin}
  \item{gazeorigin_y}{Y coordinates of the gaze origin}
  \item{gazeorigin_z}{Z coordinates of the gaze origin}
  \item{gazeDirection_x}{X coordinates of the gaze direction}
  \item{gazeDirection_y}{Y coordinates of the gaze direction}
  \item{gazeDirection_z}{Z coordinates of the gaze direction}
  \item{gazePoint_x}{X coordinates of the gaze point}
  \item{gazePoint_y}{Y coordinates of the gaze point}
  \item{gazePoint_z}{Z coordinates of the gaze point}
}
}
\usage{
gap_fill(data, max_gap_length = 75)
}
\arguments{
\item{data}{Data frame of the eye tracking data we want to process}

\item{max_gap_length}{Maximum gap length in ms which we want to fill}
}
\value{
The modified input data frame with the gaps filled and an additional column \emph{modified} which indicates if the line was modified
}
\description{
This function fills gaps in the eye tracking data by inserting a linear interpolation
between the last valid sample before the gap and the first valid sample after the gap.
}
