% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_fixations_iaoi.R
\name{merge_fixations_iaoi}
\alias{merge_fixations_iaoi}
\title{Merge Fixations from I-AOI Classifier}
\format{
Input data frame columns
\describe{
  \item{gazePoint_x}{X coordinates of the gaze point}
  \item{gazePoint_y}{Y coordinates of the gaze point}
  \item{gazePoint_z}{Z coordinates of the gaze point}
  \item{eventIndex}{Index of the event in which the gaze point lies}
  \item{eventType}{Type of the current event}
  \item{eventDuration}{Duration of the current event}
  \item{fixation_x}{X coordinate of the current fixation}
  \item{fixation_y}{Y coordinate of the current fixation}
  \item{fixation_z}{Z coordinate of the current fixation}
}
}
\usage{
merge_fixations_iaoi(data, max_time = 75)
}
\arguments{
\item{data}{Data frame of the eye tracking data we want to process}

\item{max_time}{Max time between fixations in ms}
}
\value{
The input data frame with the fixations merged
}
\description{
Merge fixations classified by \code{\link[arett]{classify_iaoi}} into a new, combined fixation when they are on the same AOI and closer to each other than the max time specified.
}
\details{
This is a custom implementation for merging fixations on the same AOI with a gap smaller than the specified time between them.
}
