% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_fixations_idt.R
\name{merge_fixations_idt}
\alias{merge_fixations_idt}
\title{Merge Fixations from I-DT Classifier}
\format{
Input data frame columns
\describe{
  \item{gazeorigin_x}{X coordinates of the gaze origin}
  \item{gazeorigin_y}{Y coordinates of the gaze origin}
  \item{gazeorigin_z}{Z coordinates of the gaze origin}
  \item{gazePoint_x}{X coordinates of the gaze point}
  \item{gazePoint_y}{Y coordinates of the gaze point}
  \item{gazePoint_z}{Z coordinates of the gaze point}
  \item{classification}{Classification of this gaze point}
  \item{eventIndex}{Index of the event in which the gaze point lies}
  \item{eventType}{Type of the current event}
  \item{eventDuration}{Duration of the current event}
  \item{fixation_x}{X coordinate of the current fixation}
  \item{fixation_y}{Y coordinate of the current fixation}
  \item{fixation_z}{Z coordinate of the current fixation}
}
}
\usage{
merge_fixations_idt(data, max_time = 75, dispersion_threshold = 1.6)
}
\arguments{
\item{data}{Data frame of the eye tracking data we want to process}

\item{max_time}{Max time between fixations in ms}

\item{dispersion_threshold}{Dispersion threshold for the combined fixation in degrees}
}
\value{
The input data frame with the fixations merged
}
\description{
Merge fixations classified by \code{\link[arett]{classify_idt}} into a new, combined fixation when they are closer to each other than the max time and the max dispersion specified is not exceeded.
}
\details{
This is a custom implementation for merging fixations on the same AOI with a gap smaller than the specified time between them.
It is intended to be used together with the I-DT algorithm described by \href{https://doi.org/10.3390/s20174956}{Llanes-Jurado et al. (2020, doi:10.3390/s20174956)} and as implemented in \code{\link[arett]{classify_idt}}.
}
