% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_fixations_ivt.R
\name{merge_fixations_ivt}
\alias{merge_fixations_ivt}
\title{Merge Fixations from I-VT Classifier}
\format{
Input data frame columns
\describe{
  \item{gazeorigin_x}{X coordinates of the gaze origin}
  \item{gazeorigin_y}{Y coordinates of the gaze origin}
  \item{gazeorigin_z}{Z coordinates of the gaze origin}
  \item{gazePoint_x}{X coordinates of the gaze point}
  \item{gazePoint_y}{Y coordinates of the gaze point}
  \item{gazePoint_z}{Z coordinates of the gaze point}
  \item{classification}{Classification of this gaze point}
  \item{eventIndex}{Index of the event in which the gaze point lies}
  \item{eventType}{Type of the current event}
  \item{eventDuration}{Duration of the current event}
  \item{fixation_x}{X coordinate of the current fixation}
  \item{fixation_y}{Y coordinate of the current fixation}
  \item{fixation_z}{Z coordinate of the current fixation}
}
}
\usage{
merge_fixations_ivt(data, max_time = 75, max_angle = 0.5)
}
\arguments{
\item{data}{Data frame of the eye tracking data we want to process}

\item{max_time}{Max time between fixations in ms}

\item{max_angle}{Max angle between fixations in degrees}
}
\value{
The input data frame with the fixations merged
}
\description{
Merge fixations classified by \code{\link[arett]{classify_ivt}} into a new, combined fixation when they are closer to each other than the max time and max angle distance specified.
}
\details{
This is an implementation of the merge fixation step as documented in the \href{https://www.tobiipro.com/siteassets/tobii-pro/learn-and-support/analyze/how-do-we-classify-eye-movements/tobii-pro-i-vt-fixation-filter.pdf}{Tobii Pro I-VT fixation filter}.
It checks each gap between fixation if it is shorter than the specified max time between fixations
and the angle between the fixations before and after the gap is smaller than specified.
If this is the case the two fixations and the gap are merged together into a new, combined fixation.

The default values for max time and max angle are defined by Tobii to be 75 ms and 0.5 degrees.
}
