% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_reduction.R
\name{noise_reduction}
\alias{noise_reduction}
\title{Reduce Noise}
\format{
Input data frame columns
\describe{
  \item{gazeorigin_x}{X coordinates of the gaze origin}
  \item{gazeorigin_y}{Y coordinates of the gaze origin}
  \item{gazeorigin_z}{Z coordinates of the gaze origin}
  \item{gazeDirection_x}{X coordinates of the gaze direction}
  \item{gazeDirection_y}{Y coordinates of the gaze direction}
  \item{gazeDirection_z}{Z coordinates of the gaze direction}
  \item{gazePoint_x}{X coordinates of the gaze point}
  \item{gazePoint_y}{Y coordinates of the gaze point}
  \item{gazePoint_z}{Z coordinates of the gaze point}
}
}
\usage{
noise_reduction(data, method = median, window_size = 3)
}
\arguments{
\item{data}{Data frame of the eye tracking data we want to process}

\item{method}{function which is given the current data window and returns the denoised value (usually either median or mean)}

\item{window_size}{Number of data points over which we want to apply the noise reduction}
}
\value{
The modified input data frame with the noise reduction applied and an additional column \emph{modified} which indicates if the line was modified
}
\description{
This function reduces noise in the eye tracking data by applying the specified method (usually either median or mean) over the specified window
}
\details{
This is an implementation of the noise reduction function as documented in the \href{https://www.tobiipro.com/siteassets/tobii-pro/learn-and-support/analyze/how-do-we-classify-eye-movements/tobii-pro-i-vt-fixation-filter.pdf}{Tobii Pro I-VT fixation filter}.
It goes through all gaze points and replaces each data point by the result of the specified method (in case of Tobii either median or mean) over all data points
in a window centered around the data point currently being replaced.

As the eye tracking rate of the HoloLens 2 is 30 Hz, a noise reduction over 3 samples would result in a window size of 1/30Hz = 100ms.
With the typical fixation duration around 200ms this would result in a window not much smaller than the fixation duration and risk
significantly influencing the data negatively. Therefore it should not be used by default.
}
